% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{ggpalette_pal}
\alias{ggpalette_pal}
\title{ggpalette_palette}
\usage{
ggpalette_pal(primary = "pink", other = "orange", direction = 1)
}
\arguments{
\item{Usage:}{ggpalette(primary = "pink",  other = "orange", direction = 1)}

\item{Arguments:}{primary    specify a manual color
other      specifies the name of the second (or other) color
direction  default value is 1; allows the user reverse the order of the palette}

\item{Value:}{n is the number of colours being picked, and returne message or warn the user if n is larger than the number of colours in the palette.}
}
\description{
ggpalette_palette
}
\section{explanation}{
 A function that takes a single argument n and returns n colors,  and let the user re-order the colors for specific values of n
}

