% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_colour_ggpalette_d}
\alias{scale_colour_ggpalette_d}
\title{Discrete ggplot palettes}
\usage{
scale_colour_ggpalette_d(
  primary = "pink",
  other = "orange",
  direction = 1,
  ...
)
}
\arguments{
\item{Usage:}{scale_colour_ggpalette_d (primary = "pink",  other = "orange", direction = 1)}

\item{Arguments:}{primary    specify a manual color
other      specifies the name of the second (or other) color
direction  default value is 1; allows the user reverse the order of the palette}
}
\description{
Discrete ggplot palettes
}
\section{explanation}{
A function that uses ggplot2::discrete_scale to allow ggplot2 to use the colour palette
}

