% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continous-scale.R
\name{scale_colour_ggpalette_seq_c}
\alias{scale_colour_ggpalette_seq_c}
\title{Generate a sequential continuous colour scale}
\usage{
scale_colour_ggpalette_seq_c(name = "pink", direction = 1, amount = 0.2, ...)
}
\arguments{
\item{Usage:}{scale_colour_ggpalette_seq_c (name = "pink", direction = 1, amount = 0.2, ...)}

\item{Arguments:}{name       defaullt color is pink; let the user select a color from ggpalette color palette
direction  default value is 1; allows the user reverse the order of the palette
amount     define the amount of lightening}
}
\description{
Generate a sequential continuous colour scale
}
\section{explanation}{
 A function that allows a user to pass a name of a colour in your palette and generate a sequential continuous colour scale
}

