% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-scale.R
\name{scale_fill_ggpalette_d}
\alias{scale_fill_ggpalette_d}
\title{For fill aesthetics in ggplot2}
\usage{
scale_fill_ggpalette_d(primary = "pink", other = "orange", direction = 1, ...)
}
\arguments{
\item{Usage:}{scale_fill_ggpalette_d (primary = "pink",  other = "orange", direction = 1)}

\item{Arguments:}{primary    specify a manual color
other      specifies the name of the second (or other) color
direction  default value is 1; allows the user reverse the order of the palette}
}
\description{
For fill aesthetics in ggplot2
}
\section{explanation}{

A function that allows your palette to be used for fill aesthetics in ggplot2
}

